LayoutTests/imported/w3c/web-platform-tests/cookies/schemeful-same-site/schemeful-iframe-subresource.tentative-expected.txt



FAIL SameSite cookies with intervening cross-scheme iframe and subresources assert_equals: `samesite_lax=0.8697851869315117` in request to `https://web-platform.test:9443`. expected (undefined) undefined but got (string) "0.8697851869315117"
